Updated where the user authentication checking happens so it can be used in all places of the app
