+++ api/add-user/src/main.rs
@@ -1,7 +1,7 @@
 use lambda_http::{service_fn, Error, IntoResponse, Request, Body};
 use sqlx::PgPool;
 use data_access::user::Users;
-use endurance_racing_planner_common::user::User;
+use endurance_racing_planner_common::User;
 use utilities::{bad_request_response, created_response};
 
 #[tokio::main]
