--- common/src/google_open_id_claims.rs
@@ -1,19 +1,0 @@
-﻿use serde::{Serialize, Deserialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct GoogleOpenIdClaims {
-    pub iss: String,
-    pub azp: String,
-    pub aud: String,
-    pub sub: String,
-    pub email: String,
-    pub email_verified: bool,
-    pub at_hash: String,
-    pub nonce: String,
-    pub name: String,
-    pub picture: String,
-    pub given_name: String,
-    pub family_name: String,
-    pub locale: String,
-    pub jti: String
-}
