--- common/src/user.rs
@@ -1,10 +1,0 @@
-﻿use serde::{Serialize, Deserialize};
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct User {
-    pub id: i32,
-    pub name: String,
-    pub email: String,
-    pub oauth_id: String
-}
